SP5: move axis primitives and rendering logic
