+++ src/ScottPlot5/ScottPlot5/Primitives/TickStyle.cs
@@ -1,4 +1,4 @@
-﻿namespace ScottPlot.Axis;
+﻿namespace ScottPlot;
 
 public struct TickStyle
 {
